refactor: 채팅방 입장 시 상대방 정보 조회 API Controller 변경 및 Swagger 그룹명 변경 및 주석처리 삭제
