+++ src/main/java/com/mjuAppSW/joA/domain/memberProfile/MemberProfileService.java
@@ -10,7 +10,16 @@ import com.mjuAppSW.joA.domain.memberProfile.dto.request.BioRequest;
 import com.mjuAppSW.joA.domain.memberProfile.dto.response.MyPageResponse;
 import com.mjuAppSW.joA.domain.memberProfile.dto.request.PictureRequest;
 import com.mjuAppSW.joA.domain.memberProfile.dto.response.SettingPageResponse;
+import com.mjuAppSW.joA.domain.memberProfile.dto.response.UserInfoResponse;
 import com.mjuAppSW.joA.domain.memberProfile.exception.InvalidS3Exception;
+import com.mjuAppSW.joA.domain.room.Room;
+import com.mjuAppSW.joA.domain.room.RoomRepository;
+import com.mjuAppSW.joA.domain.room.exception.RoomNotFoundException;
+import com.mjuAppSW.joA.domain.roomInMember.RoomInMember;
+import com.mjuAppSW.joA.domain.roomInMember.RoomInMemberRepository;
+import com.mjuAppSW.joA.domain.roomInMember.exception.RoomInMemberNotFoundException;
+
+import com.mjuAppSW.joA.domain.memberProfile.vo.UserInfoVO;
 import com.mjuAppSW.joA.domain.vote.VoteRepository;
 import com.mjuAppSW.joA.common.storage.S3Uploader;
 import com.mjuAppSW.joA.domain.memberProfile.dto.response.VotePageResponse;
@@ -27,6 +36,8 @@ public class MemberProfileService {
 
     private final HeartRepository heartRepository;
     private final VoteRepository voteRepository;
+    private final RoomRepository roomRepository;
+    private final RoomInMemberRepository roomInMemberRepository;
     private final MemberChecker memberChecker;
     private final S3Uploader s3Uploader;
 
@@ -51,6 +62,16 @@ public class MemberProfileService {
 
     public LocationPageResponse getLocationPage(Long sessionId) {
         return LocationPageResponse.of(memberChecker.findFilterBySessionId(sessionId));
+    }
+
+    public UserInfoResponse getUserInfo(Long roomId, Long memberId){
+        Room room = roomRepository.findById(roomId).orElseThrow(RoomNotFoundException::new);
+        Member member = memberChecker.findBySessionId(memberId);
+        RoomInMember roomInMember = roomInMemberRepository.findByRoomAndMember(room, member).orElseThrow(
+            RoomInMemberNotFoundException::new);
+
+        UserInfoVO userInfoVO = roomInMemberRepository.getUserInfo(roomInMember.getRoom(), roomInMember.getMember());
+        return UserInfoResponse.of(userInfoVO.getName(), userInfoVO.getUrlCode(), userInfoVO.getBio());
     }
 
     @Transactional
