+++ src/main/java/com/mjuAppSW/joA/domain/memberProfile/vo/UserInfoVO.java
@@ -1,4 +1,4 @@
-package com.mjuAppSW.joA.domain.roomInMember.vo;
+package com.mjuAppSW.joA.domain.memberProfile.vo;
 
 public interface UserInfoVO {
     String getName();
