+++ src/main/java/com/mjuAppSW/joA/domain/roomInMember/RoomInMemberApiController.java
@@ -76,17 +76,4 @@ public class RoomInMemberApiController {
         roomInMemberService.checkRoomInMember(request);
         return ResponseEntity.ok().build();
     }
-
-    @Operation(summary = "채팅방 입장시 상대방 정보 조회", description = "상대방 정보 조회 API")
-    @ApiResponses(value = {
-        @ApiResponse(responseCode = "200", description = "상대방 정보 반환"),
-        @ApiResponse(responseCode = "404", description = "M001: 사용자를 찾을 수 없습니다."),
-        @ApiResponse(responseCode = "404", description = "R003: 방을 찾을 수 없습니다."),
-        @ApiResponse(responseCode = "404", description = "RIM001: 채팅방을 찾을 수 없습니다.")
-    })
-    @GetMapping("/{roomId}/{memberId}/userinfo")
-    public ResponseEntity<SuccessResponse<UserInfoResponse>> getUserInfo(@PathVariable("roomId") Long roomId, @PathVariable("memberId") Long memberId){
-        return SuccessResponse.of(roomInMemberService.getUserInfo(roomId, memberId))
-            .asHttp(HttpStatus.OK);
-    }
 }
